refactor(channels): share one StripCompressionMarkers helper (PR #384 review)
Review flagged that ctxzipMarkerRe + stripCompressionMarkers were
byte-identical copies in slack.go and telegram.go (a drift risk). Hoist
a single exported markdown.StripCompressionMarkers into the shared
channels/markdown package that both plugins already import, and route
Slack, Telegram, and MS Teams through it. The pure-function unit test
moves to the markdown package; per-plugin tests keep the extractText /
extractLargestFile leak-guard coverage.

MS Teams' extractText had the same raw-text path and now strips markers
too — consistent coverage across every channel.

The format is still pinned separately as CompressionMarkerPrefix in
forge-core/runtime for the documented import-cycle reason; the plugins
have no such cycle, so they share.
