+++ forge-plugins/channels/slack/longresponse_test.go
@@ -13,44 +13,6 @@ import (
 	"github.com/initializ/forge/forge-core/channels"
 )
 
-func TestStripCompressionMarkers(t *testing.T) {
-	tests := []struct {
-		name, in, want string
-	}{
-		{"no marker", "plain text", "plain text"},
-		{
-			"trailing marker",
-			"Here are the issues.\n\n<<ctxzip:2cda42236849 55_lines_offloaded>>",
-			"Here are the issues.",
-		},
-		{
-			"inline marker",
-			"before <<ctxzip:abc123abc123 12_rows_offloaded>> after",
-			"before  after",
-		},
-		{
-			"multiple markers",
-			"<<ctxzip:aaaaaaaaaaaa 1_rows_offloaded>>x<<ctxzip:bbbbbbbbbbbb 2_rows_offloaded>>",
-			"x",
-		},
-		{
-			"bare marker no note",
-			"data<<ctxzip:0123456789ab>>",
-			"data",
-		},
-		// A "<<ctxzip:" that is not a well-formed marker is left alone rather
-		// than eating the rest of the string.
-		{"malformed left intact", "text <<ctxzip:not-hex blah", "text <<ctxzip:not-hex blah"},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := stripCompressionMarkers(tt.in); got != tt.want {
-				t.Errorf("stripCompressionMarkers(%q) = %q, want %q", tt.in, got, tt.want)
-			}
-		})
-	}
-}
-
 // TestExtractText_StripsCompressionMarker pins that a leaked ctxzip marker
 // never reaches a Slack channel through the text path.
 func TestExtractText_StripsCompressionMarker(t *testing.T) {
