+++ forge-plugins/channels/slack/slack.go
@@ -9,7 +9,6 @@ import (
 	"io"
 	"net/http"
 	"net/url"
-	"regexp"
 	"slices"
 	"strconv"
 	"strings"
@@ -1002,23 +1001,6 @@ func unwrapJSONContent(text string) string {
 	return text
 }
 
-// ctxzipMarkerRe matches an internal context-compression pointer
-// ("<<ctxzip:HASH note>>") that ctxzip leaves in place of offloaded content.
-// Mirrors github.com/initializ/ctxzip/ccr's marker format.
-var ctxzipMarkerRe = regexp.MustCompile(`<<ctxzip:[0-9a-f]{12,64}(?:[ ,][^>]*)?>>`)
-
-// stripCompressionMarkers removes any ctxzip markers that leaked into
-// channel-bound text. These are internal artifacts of context compression; a
-// user must never see them. The model is expected to expand offloaded content
-// it needs before answering, so a marker reaching a channel is a leak — drop
-// it rather than surface a dangling pointer.
-func stripCompressionMarkers(s string) string {
-	if !strings.Contains(s, "<<ctxzip:") {
-		return s
-	}
-	return strings.TrimSpace(ctxzipMarkerRe.ReplaceAllString(s, ""))
-}
-
 // extractText concatenates all text parts from an A2A message.
 func extractText(msg *a2a.Message) string {
 	if msg == nil {
@@ -1033,7 +1015,7 @@ func extractText(msg *a2a.Message) string {
 			text += unwrapJSONContent(p.Text)
 		}
 	}
-	text = stripCompressionMarkers(text)
+	text = markdown.StripCompressionMarkers(text)
 	if text == "" {
 		text = "(no text response)"
 	}
@@ -1057,7 +1039,7 @@ func extractLargestFile(msg *a2a.Message) (content, filename string) {
 			if strings.HasSuffix(p.File.Name, ".md") {
 				raw = unwrapJSONContent(raw)
 			}
-			content = stripCompressionMarkers(raw)
+			content = markdown.StripCompressionMarkers(raw)
 			filename = p.File.Name
 		}
 	}
