+++ forge-plugins/channels/telegram/marker_test.go
@@ -7,37 +7,6 @@ import (
 	"github.com/initializ/forge/forge-core/a2a"
 )
 
-func TestStripCompressionMarkers(t *testing.T) {
-	tests := []struct {
-		name, in, want string
-	}{
-		{"no marker", "plain text", "plain text"},
-		{
-			"trailing marker",
-			"Here are the issues.\n\n<<ctxzip:2cda42236849 55_lines_offloaded>>",
-			"Here are the issues.",
-		},
-		{
-			"inline marker",
-			"before <<ctxzip:abc123abc123 12_rows_offloaded>> after",
-			"before  after",
-		},
-		{
-			"bare marker no note",
-			"data<<ctxzip:0123456789ab>>",
-			"data",
-		},
-		{"malformed left intact", "text <<ctxzip:not-hex blah", "text <<ctxzip:not-hex blah"},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := stripCompressionMarkers(tt.in); got != tt.want {
-				t.Errorf("stripCompressionMarkers(%q) = %q, want %q", tt.in, got, tt.want)
-			}
-		})
-	}
-}
-
 // TestExtractText_StripsCompressionMarker pins that a leaked ctxzip marker
 // never reaches a Telegram chat through the text path.
 func TestExtractText_StripsCompressionMarker(t *testing.T) {
